features-rlgym-rocket-league/pyproject.toml
[project]
name = "features-rlgym-rocket-league"
version = "0.1.0"
dependencies = ["features-rlgym-api"]

[tool.setuptools.packages.find]
where = ["src"]